+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Change Log
 
+### v1.2.0 (2022-04-18)
+- Added capability to set value
+
 ### v1.1.0 (2022-04-12)
 - Changed bundler to rollup; better module generation, only generates browser module
 - Disabled behaviour; setting the element as disabled now works as expected